simplify Exporter|Importer main bases
it's simpler as it uses less concepts, mostly just functions calls, instead of inheritance and App, which bring DelayedInit and the order of execution is getting complicated

following Michal's suggestion in https://github.com/ShiftLeftSecurity/overflowdb/pull/300#pullrequestreview-963053877
